backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
psycopg2-binary==2.9.11
pydantic==2.11.1
openai==1.0.0
httpx==0.27.2
faiss-cpu==1.7.4
numpy==1.26.4
pandas==2.2.3
